requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine learning libraries
scikit-learn>=1.1.0
tensorflow>=2.10.0
torch>=1.12.0
transformers>=4.20.0

# Financial data libraries
yfinance>=0.1.70
alpha-vantage>=2.3.1
pandas-datareader>=0.10.0

# Sentiment analysis libraries
vaderSentiment>=3.3.2
textblob>=0.17.1
nltk>=3.7
spacy>=3.4.0

# Web scraping and API libraries
requests>=2.28.0
beautifulsoup4>=4.11.0
praw>=7.6.0
tweepy>=4.10.0

# Visualization libraries
plotly>=5.10.0
altair>=4.2.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.64.0
joblib>=1.1.0

# Optional: For advanced features
# streamlit>=1.20.0  # Uncomment if using Streamlit
# dash>=2.6.0        # Uncomment if using Dash 